[ticket-list] Existing ticket is now loaded at startup
To make it work, the initialization of the compat layer has to be moved
to the closest place where this was done previously: the PreLoadUnpaidOrdersHook.
In addition, the initialization has to be done in two steps: first we sync the state ticket
to a dummy OB.Model.Order object and then we copy the whole ticket to model.receipt using the
clearWith function. This way the whole ticket will be copied in one transaction and all events triggered
will work properly. If you try to sync the state ticket to the model.receipt object directly, events will
be triggered over incomplete versions of the Ticket and will fail.