Fixes ISSUE-44683: Sets correct DateTime in window grid exported CSV file
DateTime in export to CSV was wrong, applying local time zone several times instead of once.

It has been fixed by using a function that doesn't try to transform to
UTC in DateTimeUIDefinition, convertToClassicStringInLocalTime. This
function formats the date as is, using the timezone received without
prior conversion to UTC.

Fixes regression introduced in commit
4ec715bfa9c6bc0ab7b3e6f644ec05ef4998c721