Fixes ISSUE-44690: Format date using provided timezone in OBBindings#formatDate
Previously as a side-effect date was being converted to UTC wrongly in
OBBindings formatDate function.
This fixes it by formatting using the local timezone as was previously
done prior to the regression commit. Fixes display of date in some
parameter windows that were relying on it (e.g. Costing Rules).

Fixes regression introduced in commit 4ec715bf