[quartz-cluster] Allow unscheduling jobs on all instances when possible
Scheduling jobs is now permitted even when no-execute policy is applied
to instance of cluster, if and only if there's a scheduler in the
cluster that can execute jobs.

This applies the same logic for unscheduling jobs, allowing to
unschedule jobs from any instance in the cluster if and only if there's
an active scheduler in any of the instances of the cluster.