[ticket-list] Don't execute calculateReceipt when propagating changes state->backbone
  The calculateReceipt function should not be executed when propagating
state->backbone changes, as its logic is now implemented using a Ticket
model hook.

  This was supposed to be already implemented but it was a bug in the
calculateReceipt function because it was not checking the correct
variable to skip its calculation. Now this is fixed.

  Besides we are also removing code that was setting the (incorrect)
flag to avoid the calculateReceipt execution after firing some actions.
Now this code is no longer needed as the platform takes care of setting
the (correct) flag which now is correctly taken into account inside the
calculateReceipt function.