[quartz-cluster] Set misfireThreshold back to 60 seconds and fix MisfireTests
Misfire threshold property allows processes to miss their starting time
by 60 seconds and don't be considered misfired. Previously it was set to
60 seconds but because there was some problem with wait times in
MisfirePolicyTest it was changed to 1 seconds.

A workaround has been set in place, tests that have to wait for the
MisfireHandler will wait the time in property previous to checking if it
was misfired or not.