[TRL] Reduced Translations
This development allows to exclude elements to be translated from Openbravo.
To do so, a new flag has been created for each Menu entry. If the flag
is not enabled, the menu entry can be excluded while creating the translation.

Additionally, in the Window to Export the Translations a new flag has been created
to allow to generate the full or the minimized version of the translation.

The development includes:
* New flag for the Menu entries
* Exclude some menu entries from the minimized translation
* New flag in Export Translations Window
* Code to exclude the Translations based on this flag
* Also:
  - Exclude Translations from non active records
  - Exclude Translations from jrxml files
  - Exclude Translations form AD_Reference Table