fixes BUG-44625: update xercesImpl library
  Update xercesImpl library to version 2.12.1. This version still
is not compatible with JDK11 as it already has a split package problem
with the org.w3c.dom.html. Note that there is a known issue[1] already
reported for this library to fix this.

  For this reason, apart from updating this library to the latest
version we are fixing the JDK11 split package problem by patching the
.jar library to remove the org.w3c.dom.html package which indeed only
contained a single class: HTMLDOMImplementation.

  [1] https://issues.apache.org/jira/browse/XERCESJ-1689