Fixes ISSUE-44963: Make quartz.properties a template file
quartz.properties is removed and a template file is created as it makes
more sense. This file has also been added to be copied in the setup ant
task.

In quartz.properties.template there are 2 different configurations, one
for non-clustered environment(default) and one commented-out for clustered environment.

A new module script has been created to handle recreation of this file
on previous versions.