FIXES ISSUE-44938. Fixes problem when persisting information in Costing process
Before, when finding a particular condtion, a return statement was executed,
finishing the loop but also finalizing the execution for the rest of the
code of the method, which included some final validations + a flush and a clear.
Due to this fact, the changes were not persisted in the Database.

This change allows to break the execution of the loop, but keep executing the
final validations + flush and clear as it was intended.

Also, for this case, a new boolean variable named forceExit has been used
instead of the more common 'break' sentence to finish the loop.
It has been done because the code inside the loop is too big to understand
everything that is happening, so it is better to explicitly explain that
the loop can be prematurely finished, instead of doing it through a buried
return statement.