Fixes ISSUE-44921: On hiding or showing grid column, don't refresh on lazy filtering
Previously when hiding or showing a column in grid view a request was
made to refresh the content of the grid, resulting in lazy filtering
being ignored.
This fix makes sure that showing or hiding a column in a grid view when
lazyFiltering is on will not trigger a refresh of the grid view content.