Fixes ISSUE-45006: Stabilize MisfirePolicyTest when clustered
MisfirePolicyTest checkMisfirePolicyWithSecondlySchedule was not waiting
enough for the execution of the expected 2 jobs. MisfireHandler checks every
second for missed jobs to execute. A lower time for the misfireHandler has been set, 500ms
and delay has been set to 3 seconds.
Also because of this, a race condition was happening when using
Dal transaction to access Lock table by quartz, it should not happen as
misfireHandler has plenty time to check and execute jobs now.