Fixes ISSUE-45004: Remove obsolete org.openbravo.client.kernel/docs
Those docs were previously generated from javascript with
code-illuminated library. The syntax is very different to the jsdoc used
currently. So it is removed as obsolete, and both library and license
for it are also removed.

Template docs.js.ftl and class DocumentationComponent are also removed
because they're not needed anymore by the module.