Fixes ISSUE-45089: Remove leak by destroying previous loading pane always
A leak was happening when opening the same tab 2 or more times without
closing it. To fix this, the old pane should be destroyed when the window is
opened and replacing the old tab.

To achieve this, pane.isLoadingTab is used to check if the pane we are
switching to, is also loading and destroy the previous pane in this
case.