Fixes ISSUE-45165: Fix inaccessible Popups(grayed) after long usage without refreshing
Popups sometimes appeared behind a mask instead of in front of
everything else. This was happening because of an overflow on z-index
ranges used by smartclient.

To solve this, one option is to reset those z-index variables, but
that's not ok on a working environment without refreshing or
logout/login of the user. z-index range for smartclient has been
increased, putting all popups more in front, so it doesn't get
overflowed by the next z-index variable.