Fixes ISSUE-45182: AddProduct action can't be executed from new UI
  Some of the objects referenced in AddProduct are not available in the
new UI which is preventing the action to be executed properly.

  Changes applied:

  - Now the services filter implementation is placed inside
OB.App.StandardFilters.Services which is indeed available in the new UI.
And "old" OB.UTIL.servicesFilter makes use of it.

  - In the StockChecker class if OB.OBPOSPointOfSale is not available
(when we are in new UI) then we directly return true. Note that this should
be replaced in the future with a correct implementation of the stock check.