Related to ISSUE-45006: Stabilize MisfirePolicyTest tests in clustered mode
Tests in MisfirePolicyTest class were unstable in CI and as such, some
parts have been implemented with this in mind. scheduleJob method will
try to schedule a job and if it fails, it will retry after 1 second,
until it reaches maximum time of 10 seconds.

This will allow to handle some of the race-conditions that were
happening in CI in this issue.

Also, now those tests will test that the trigger has correctly misfired,
that was previously implied, but now it is actually tested using a
TriggerListener. Also waiting some time if the misfire was not
detected on the first 500ms. The maximum wait time is defined in
MAX_WAIT constant.