[jdk11] Remove JDK8 conditional for -html5 tag on JavaDoc generation
From JDK10 forwards, it is necessary to specify the html version on
generation. A conditional was set in place to add it only if version was
not Java 1.8. This conditional is removed here, as it is not necessary
anymore, -html5 tag will be specified always.

Also some extra fixes to JavaDoc generation have been added:
- Set source to use Java 11 instead of 8
- Set the correct link to the JavaSE 11 documentation, previously was a
deprecated link.