[jsengine] Remove yui-compressor and older rhino libraries
YUI-compressor is a library that was used to compress old 2.5 javascript
and css in /web directory. It was not used properly, as it didn't
compress for example ltr and rtl skin folders.

As rhino has been reintroduced in this project, older versions of rhino
are removed, those include:
ob-rhino-1.6R7.jar
js.jar
yuicompressor-2.4.2.jar (This contains an embedded version of rhino)

Build has been updated to remove the ant tasks regarding this
yuicompress library. Modern JavaScript and CSS is unaffected, as it is
minified and compressed using other tools.