Fixes ISSUE-45301: Record can be saved with missing mandatory field data
If a mandatory field in a collapsed field group had no value or invalid
values, the record could be saved by the auto-save when changing focus
to a child tab.

The approach taken here is to also validate hiddenFields that are not
present because of being in a collapsed field group.

form.validate method can be found in smartclient DynamicForm and it
accepts as first argument a boolean indicating to validate or no hidden
fields, it is now set to true on ob-view-form, fixing this issue.