[jsengine] Fix return "undefined" by Rhino when it should be null
Rhino differs from Nashorn in that it returns "undefined" when the
result of the evaluation is such, for example when a property or method
is not defined. Nashorn was previously returning null in those cases.

Because of this, now we transform "undefined" to null after script
evaluation.

Also, some extra tests have been added in OBBindings to test these
cases.