Fixes ISSUE-45430: Improve unbalanced adminMode warnings in OBBaseTest.
OBBaseTest did already check if a test-case left adminMode enabled
instead of handling it properly.
However that detection was not integration with the existing OBContext
code for doing the same logic at the end of an http-request.

Reuse that existing code allows to reuse the debugging ADMIN_TRACE_SIZE
logic from OBContext also to debug Junit test issues.