Fixes ISSUE-43275: On new table, AD_Org column name is wrong after sync terminology
Once a new table is created and sync terminology executed,
the AD_Org element changes to "Invoice Organization" instead of
"Organization". This is done because there are several elements with the
same column name in AD_ELEMENT table and internally the Max
AD_Element_id is being used in the synchronization process.

To avoid this, the main columns are being set to their respective main
elements manually, this only affects newly created tables and has no
effect on existing tables.