Fixes ISSUE-45472: Grid Horizontal scroll on RTL is not working with Chrome >= 85
Horizontal scroll on smartclient grid is failing on Chrome version 85+,
this is due to the modification of how Chrome implements horizontal
scrolling, it has been modified to follow similar standards to mozilla,
which previously it was not.

This required porting one of the patches from smartclient version 12.1,
which fixes this issue.