Fixes ISSUE-45443: Lines selected after sort Payment Proposal Payments are not included
Payment Proposal Expected Payments Pick&Execute window was losing the
Payment field value when a filter was applied or a column was sorted.
This was due to the record being modified but not the editValues from
the grid. A refactor has been made on the onSelect event handler to fix
this and use grid.setEditValue instead of the old record.value = ...

Also, a validator was also modifying values from the record, to avoid
this, the logic that modified the payment + difference has been
extracted to an onChange event handler.