fixed BUG-45621: Returns P&E shows product's name instead of its identifer
The grid of Pick Lines process definition used from Return to Vendor window,
displayed product's name instead of its identifer. The AD reference of
this column was Table which does not display identifer but a single
column.

This column's reference has been changed to TableDir which displays the
whole identifier.