Fixes ISSUE-45803: MessageBar shows on top of all windows when tab is closed before ActionHandler
A red/green message bar appears on top if a tab is closed before a
process ActionHandler had time to return the results. This was due to
how the context is being checked in getTabMessage method, an wrong
clientContext.messageBar was being used without checking if it still
exists in the current window.

This fixes it by checking if the window contains the saved
clientContext element and it's the same as the one who called the
handler.