[react] add svg mock mapper for jest tests to avoid object proxies in tests
Original mapping for .svg was identity-obj-proxy, which has a problem
when dealing with jest tests, sometimes the value of the object proxy
will not be delivered and an error is shown, because the object is
passed to methods instead of the result, which should be the resolved
path to that object.

To avoid this, a mapper fileMock.js file has been created, which exports
the .svg files using module.exports, this fixes this problem.