Fixes ISSUE-45830: Fixed error on Fix Backadated Transaction process.
In some cases was possible that transaction ordering without having into account ordering by product, caused that transactions be wrongly considered as backdated.

To fix this issue the transaction movement dates is compared on same product transactions. The getTransactions method was changed to sort the transactions also by product and always reset the comparison date of the last transaction for the product when the product changes on the loop.