Fixes ISSUE-45855. Remove calling DatabaseValidor from package.module
ant package.module runs validations for the module before creating the
obx file.
That has 2 parts:
a.) ModuleValidator
b.) DatabaseValidator

The DatabaseValidator checks are also run during export.database (for
the module being exported). So with the typical sequence of
export.database and then package.module same topics are checks twice.

As the code in this task for DatabaseValidator or blocking removal of
lots of otherwise obsolete code inside dbsourcemanager stop calling the
b.) from ant package.module and rely on people using the WARNings
already shown during export.database.