Fixes ISSUE-45853: All data is shown by default in subtabs with Lazy filtering configured
Previously when switching to a subtab of a subtab, lazy filtering was
not being used and a refresh of the data was being executed. To avoid
this, when switching to a tab with a grid that is not yet visible, only
force a refresh if lazy filtering is not configured for that grid.