fixed ISSUE-45890: ClassLoaderTest fails if a Tomcat filter is added
Tomcat filters (ie. org.apache.catalina.filters.ExpiresFilter) are not
included in the classpath when executing ClassLoaderTesst from CLI. This
makes the test to fail they are included.

Fixed by allowing org.apache.catalina.filters.* classes as filters, note
they are skipped, so it is not guaranteed they actually exist.