Fixed BUG-45909: Remove trigger and improve unique constraint
The trigger has been replaced by an unique index that includes
the currencies and a coalesce between all the possibles document ids
that can be involved in a conversion rate document. Note that at least
one of them must be not-null.

This new unique index is the improved version of the existing unique constraint
(C_CONVERSIONRATEDOC_ONCE) that has been removed as it wasn't working properly.
With the new unique index there is no need to have the C_CONVERSION_RATE_DOC_TRG2
trigger as it was implementing the same functionality that the new unique index is
doing.

Changed existing error message key to match the unique constraint name.

Transform some indexes to partial index after trigger removal, as there is no real
advantage of having those indexes as full indexes anymore.