Related to BUG-45909: JUnit tests stabilization
In Oracle the tests were working fine because the transaction was kept alive,
so the code was able to properly remove the changes in the database in @After.

However, PG has a different transaction management and the transaction was aborted
in the moment the exception was uncaught.

To fix this we make sure the last executed test is the one that must fail, and in
the @After we just rollback the transaction