Fixed BUG-45641: Overpayment action selector appears by scrolling down
Embed call to AddPaymentDisplayLogicActionHandler inside a fireOnPause() to avoid race condition.

It was observed that before the fix there were 4 calls to this handler, where usually the output was:
Y, Y, Y, N, which is the correct, as the last response is N. However, there were cases where the last
response was Y, making the field wrongly visible.

After this changeset it has been observed that just one call is done, with output N.