Fixes ISSUE-45681: Values not updated after saving fails in a P&E grid that allows additions
This issue is due to showInlineEditor method in pick and execute
returning early and not executing completely when a mandatory field has
not been completed in the currently edited line.
To avoid it, the inline editor will open if the current line in a pne
grid still has some mandatory fields empty.