Fixes FR-46008: Jest tests don't fail when console.warn or console.error appear in a test
This is due to how jest works, it doesn't directly check if a
console.warn or a console.error has been thrown during a test execution.
To check this, console.error and console.warn functions are modified in
jest.setup.js so it handles and fails correctly the test that showed
some console.error/warn.