[npm-install-script] Updates npm install script to ensure versions defined in lock files are used
The new script does not use lerna bootstrap to install the npm dependencies. It was not suitable for our usecase, where:
- we could not use hoisting because that is not compatible with installing modules with npm ci, not being able to lock dependencies properly
- if dependencies are not hoisted then we need to create symlinks between the node_modules folders of Openbravo dependencies to prevent having
to define redundant dependencies, i.e. if org.openbravo.pos2 depends on org.openbravo.core2 and the latter includes a npm dependency to react,
we need the node_modules folder of the former to include the react package too. We don't want to redefine the dependency in the package.json of
org.openbravo.pos2 so we prefer to include symlinks instead. We did not found a way to do this with lerna, so we are defining our own install script

The code that given a module returns the list of npm dependencies defined in modules that it depends on has been extracted to a NpmDependencyReader
class to be reusable.

Lerna dependency will be removed in a later changeset if we determine it is no longer needed.