[npm-install-script] Link modules to base node folder, instead of each dependant module node folder
Previous approach had a performance impact, when updating any watched file the time it took webpack to recompile
the sources went from ~5 seconds to ~30 seconds. Now instead of linking the npm dependencies from one module to another,
the dependenecies are linked from the openbravo root node_modules folder to their original location. This preventes
the performance problem and on top of it fixes the Storyshots tests, that relied on the modules being hoisted.

Because first-level dependencies are now hoisted, the install modules script now includes a validation that checks if the
same dependency is included in several openbravo modules. If two modules include the same dependency with the same version a
warning message will be shown, if the same dependency is defined with different versions an error will be shown and the module
installation will be canceled