[ALLG] Added Characteristic type to M_Characteristic
It allows the possibility to define types of product characteristics that might
be interesting for a concrete business. For example, in restaurants a product
characteristic type might be defined as "Allergens". This definition might
allow to develop advanced features.

By default the type is empty (null), which means the product characteristic
doesn't belong to a concrete type.

Using a List instead of a flag also avoids performance penalty when creating the
new column in existing instances.