Fixes ISSUE-45931: Create Fields button in Window -> tabs not working (0 fields created)
This issue was due to a modification on the selection query to detect
which fields should be added, from the columns of the underlying table.
A new join was added with table user_tab_column, but it was incorrect,
as one part of it was always in upper case, while the other part was not.

To fix it, an UPPER() function has been added to the corresponding where
clauses that required it.

Fixes regression introduced by commit 809d27effffdab2849401ef6ae46b0887286d4f3