Fixes ISSUE-46154: eslint ignores web-jspack if org.openbravo.core2 is not installed
This is needed because files under web-jspack may potentially reference components from core2 and pos, and
those dependencies are not mandatory. Now all files under web-jspack are ignored if org.openbravo.core2
is not installed.

Also adds -no-error-on-unmatched-pattern to prevent errors if there are no files to lint after excluding
those under the web-jspack folder.