Related to ISSUE-46178: Code review improvements
Removed setFilterOnReadableOrganization(false) because it has no sense. Note that
just after this, the flow gets the list of editable organizations using:
  Utility.getContext(conn, vars, "#User_Org", "LoginHandler")
and it iterates over them (and their parent tree) to get the ledger configuration.

The ledger configuration must be in a readable organization (either in an editable
organization or in a parent editable/readable organization), so the filter by
readable organization should be in.

Note that ideally we should filter by the editable organizations and their parents,
which is what later on is going to be used, however the purpose of this fix is to
avoid performance penalty, so having a fast (although inaccurate query in some corner
cases) is a good approach.

Apart from that the method has been renamed to be more clear.