Fixes ISSUE-46245: sourcedata and sampledata of uninstalled optional modules is not removed
Now dual modules can have sourcedata and sampledata of tables that belong to optional modules (i.e. org.openbravo.core2, OBC2 dbprefix).

When running update.database the sourcedata will just not be installed (no change needed for this), and then when running
export.database, the OBC2 source data will not be removed even though it was missing from the database.

When importing the sampledata, those of tables that belong to uninstalled optional modules will not be imported, and when running
export.sample.data afterwards, those files will not be deleted.

The change has been done so that it is easy to support more optional modules (i.e. org.openbravo.pos2).