Fixes ISSUE-46377: Fix create.database with PostgreSQL 14.
Some very old code to check & create the pl/pgsql language is running in
prescript and is how having syntax issue with upcoming PG14.

However since PG 9.0 the pl/pgsql language is installed by default (and
as an extension) so that code should not be needed anway (as we require
PG 10 since 20Q1 release already).

Remove that code as obsolete which indirectly fixes the PG 14 problem.