[color-picker] Change color picker to be a selector reference instead of table
This allows it to be used in Process Windows as a parameter, which is
not possible when using a sub-reference of type table.

As part of this, a modification in OBViewParameterHandler has been
introduced that allows custom UIDefinitions from sub-references to be
used instead of their parent reference UIDefinition when present.