Fixed ISSUE-46501: Requisition line qty min value allows > 0
The min value of any quantity column has been changed from 1 to 1E-20, which
easily allows to save a value greater than 0 with enough precission.

Note that the NumericPropertyValidator.validate() is in charge of validating the number.
It uses the BigDecimal(String) constructor, where the param is "1E-20", which is a valid
string for this constructor.