fixed BUG-46521: ant compile.test does not work incrementally
Whenever compile.test task was executed all test classes were compiled
instead of doing an incremental compilation of the only the ones whose
source files changed since last compilation.

This was caused by an incorrect path definition of tests in core module,
sources are within src-test/src directory but pathelement was declared
as src-test only, this makes ant not to be able to detect changes and to
always compile all the files. See [1].

[1] https://ant.apache.org/faq.html#always-recompiles