Fixes ISSUE-47145: Import entries can be permanently left unprocessed
The problem was that it was possible for ImportEntryProcessor to cache a Runnable assuming that it had been
properly submitted to a executorService, when in reality the submission had failed (i.e. because the queue
of the executorService was at maximum capacity).

This has been fixed by ensuring that the runnable is not cached until ensuring that it has been properly
submitted

The visibility of the submitRunnable method has been changed to package, because there is only need for
ImportEntryProcessor to have access to it.