fixes BUG-46930: tabs hidden by display logic are accessible from menu
When tabs are hidden by display logic and the whole tab set does not fit
in the window, they are still present and accesible from the drop down tab
menu selector.

SmartClient shows as disabled those tabs which are marked as disabled
but it does not provide a way not to display some tabs in the menu (see
SC's TabSet.showTabPickerMenu). Hidden tabs will be marked as disabled
in the menu.