Fixes ISSUE-37498. Remove deprecated EMAil class.
That old class was make obsolete and deprecated in MP9.
However code in AlertProcess was left using it as a fallback.

However the modulescript UpgradeToPoc email (added back then) was
auto-upgrading the config from ad_client to c_poc_configuration already.

That made this fallback code kind of obsolete already, remove it
completely.
That allows to delete the obsolete files.