Fixes BUG-46886: Fix already shipped quantity in RTVS
The already shipped quantity was only be calculated correctly if preference "Enable UOM Management" was configured, in other cases process was not taking into account already shipped quantities.
To fix this, if the preference is not configured any conversion is required so the already shipped qty will be the sum of movement quantities in all the lines.