fixes BUG-47236: "changes pending" displayed after MMC module install + rebuild
After following the steps described in the doc to install a module:
 - Install module from MMC
 - ant update.database
 - ant compile.complete.deploy

the module remains in status Pending (P).

Status is updated to Applied (A) on install.source and smartbuild.

This fix moves the status update from smartbuild to update.database
after applying modules (importing their reference data) where this flag is
used. In fact, this is more correct than what was done before as
smartbuild without datbase update could lead modules in an incorrect
status.