Fixes BUG-47045: Set the order reference to an invoice header when the Order is updated.
When the C_Order_Id is filled (insert/update), the corresponding reference no. is inserted
unless the user did manually enter any other value, which in this case it's kept.

When the C_Order_Id is removed, the reference no. is removed too only if it's exactly the
same value as the order being removed, otherwise it's kept (which means the user manually
entered any other reference here).

The logic is managed in a trigger, so any invoicing process will get the same logic.