Fixes ISSUE-47309: Saving record from form view sometimes fails if grid does not have all data in cache
The problem was the following:
- Start with a grid where not all data is cached (i.e. Sales Order after cleaning the filter)
- Use a grid sort so that the new record would end up after the last one shown in the grid (i.e. documentno asc)
- Add a record in form view. The record is initially added to the grid in the first row, but the filter is applied right after
- Because the new record would not be in the first page, the grid would move the viewport to display it and do datasource requests to obtain the proper page
- These unexpected datasource requests result in grid selections not storing the new data properly

To fix this we follow a similar approach to what is done when adding a new row to the grid while not having all the data cached [1]. The new row is stored in the first
position, regardless of the filter. Now when storing a new record in form view, the grid filter is removed if not all the data is cached, and as a result the new record
will be stored in the first position, and the grid selection will work as expected.

[1] https://gitlab.com/openbravo/product/openbravo/-/blob/master/modules/org.openbravo.client.application/web/org.openbravo.client.application/js/grid/ob-view-grid.js#L3663